rename `sharing` to `core`
