+++ configs/rover_configs.yaml
@@ -15,5 +15,5 @@
 # limitations under the License.
 #
 
-sharing:
+core:
   active: ${ROVER_SHARING_ACTIVE:true}
